email: allow to handle any (supported) event
By naming the templates after the Event.name and generating the email
message independent of the Event type we can support any Event just by
adding a template for it.

In addition, since the signal handlers are going to all be the same, we
can make a generic signal handler factory.