recipients: put method config to a config field
Right now Recipient has an email field. However adding notification
methods would require adding new fields to the Recipient class. This
could get out of hand. Therefore this change adds a  config field which
is a dict[str, str]. They key will be the method name and the value is
the config for that method.