give events the same names as their signals
From here "events" and "signals" are basically the same thing, so
"build_pulled" has been renamed to "postpull" and "build_published"
renamed to "published".

This is a breaking change.

The motivation for this change is that I want to be able to send
notifications for signals that are not "built in" (signals coming from
other plugins). So this needs to be dynamic which means there shouldn't
be any hard-coded mappings.

As such since events are dynamic a Settings attribute has been created,
`EVENTS`. This setting tells gbp-notifications which signals to listen
for. The default is "postpull" and "published".